+++ yshphys/yshphys/Sphere.cpp
@@ -14,6 +14,6 @@ Sphere::~Sphere()
 
 dVec3 Sphere::Support(const dVec3& v) const
 {
-	float k = m_radius / sqrtf(v.Dot(v));
+	double k = m_radius / sqrt(v.Dot(v));
 	return m_pos + v.Scale(k);
 }
